Updated ingestion tests
